+++ src/components/common/alarms/AlarmSet.tsx
@@ -1,8 +1,8 @@
 import getCookies from '@/src/lib/getCookies';
-import AlarmIcon from './AlarmIcon';
+import AlarmContainer from './AlarmContainer';
 
 export default function AlarmSet() {
   const { userId, token, userType } = getCookies();
 
-  return <AlarmIcon userId={userId} token={token} userType={userType} />;
+  return <AlarmContainer userId={userId} token={token} userType={userType} />;
 }
